requirements.txt
# Core Framework
streamlit>=1.28.0
numpy>=1.24.0

# Vector Store (Optional but recommended)
faiss-cpu>=1.7.4  # Use faiss-gpu for GPU support

# LLM Integration
boto3>=1.28.0  # For AWS Bedrock
anthropic>=0.21.0  # For Anthropic API

# Code Processing
gitpython>=3.1.0  # Alternative to subprocess for git operations

# Data Processing
pandas>=2.0.0